+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add support for the ECLiPSe backend on Windows
 * Add test results from running the "Logtalk: "Run Project Testers" command to the Test Explorer
 * Add `logtalk.tests.createAllureReport` boolean setting to generate Allure reports after running tests
+* Add support for skipping and un-skipping tests using the test item context menu
 * Automatically recompile code in debug/normal mode when starting/stopping debugging
 * Add support for the debug toolbar with additional buttons for common port commands
 * Add "Help: Logtalk Handbook" command to open the Logtalk Handbook using Live Preview if installed, or the default browser otherwise